requirements.txt
openai
streamlit
requests
pandas
pandas_ta
schedule
plotly
numpy
pyupbit
python-dotenv